Move circuit code to a new file
